ref(redis): Address review comments.
* moved the redis tests into the redis crate
* added serde-yaml in relay-redis crate as dev deps
* keep docs properly formatted
* removed the option from the config which should not be configured
