+++ relay-config/src/config.rs
@@ -2068,82 +2068,4 @@ cache:
             Err(_)
         ));
     }
-
-    #[test]
-    fn test_redis_single_opts() {
-        let yaml = r###"
-processing:
-    enabled: true
-    kafka_config:
-    - { name: "bootstrap.servers", value: "127.0.0.1:9092" }
-
-    redis:
-        server: "redis://127.0.0.1:6379"
-        max_connections: 42
-"###;
-
-        let config: ConfigValues = serde_yaml::from_str(yaml)
-            .expect("Parsed processing redis config: single with options");
-        let redis = config.processing.redis.expect("Redis config: single");
-
-        match redis {
-            RedisConfig::SingleWithOpts { server, options } => {
-                assert_eq!(options.max_connections, 42);
-                assert!(!options.test_on_check_out);
-                assert_eq!(server, "redis://127.0.0.1:6379");
-            }
-            e => panic!("Expected RedisConfig::SingleWithOpts but got {:?}", e),
-        }
-    }
-
-    #[test]
-    fn test_redis_single_opts_default() {
-        let yaml = r###"
-processing:
-    enabled: true
-    kafka_config:
-    - { name: "bootstrap.servers", value: "127.0.0.1:9092" }
-
-    redis:
-        server: "redis://127.0.0.1:6379"
-"###;
-
-        let config: ConfigValues = serde_yaml::from_str(yaml)
-            .expect("Parsed processing redis config: single with options");
-        let redis = config.processing.redis.expect("Redis config: single");
-
-        match redis {
-            RedisConfig::SingleWithOpts { options, .. } => {
-                // check if all the defaults are correctly set
-                assert_eq!(options.max_connections, 24);
-                assert!(!options.test_on_check_out);
-            }
-            e => panic!("Expected RedisConfig::SingleWithOpts but got {:?}", e),
-        }
-    }
-
-    // To make sure that we have backwards compatibility and still support the redis configuration
-    // when the single `redis://...` address is provided
-    #[test]
-    fn test_redis_single() {
-        let yaml = r###"
-processing:
-    enabled: true
-    kafka_config:
-    - { name: "bootstrap.servers", value: "127.0.0.1:9092" }
-
-    redis: "redis://127.0.0.1:6379"
-"###;
-
-        let config: ConfigValues = serde_yaml::from_str(yaml)
-            .expect("Parsed processing redis config: single with options");
-        let redis = config.processing.redis.expect("Redis config: single");
-
-        match redis {
-            RedisConfig::Single(server) => {
-                assert_eq!(server, "redis://127.0.0.1:6379");
-            }
-            e => panic!("Expected RedisConfig::Single but got {:?}", e),
-        }
-    }
 }
