+++ relay-redis/Cargo.toml
@@ -18,3 +18,6 @@ serde = { version = "1.0.114", features = ["derive"] }
 [features]
 default = []
 impl = ["r2d2", "redis"]
+
+[dev-dependencies]
+serde_yaml = "0.8.13"
