+++ relay-redis/src/config.rs
@@ -4,23 +4,18 @@ const fn default_max_connections() -> u32 {
     24
 }
 
-/// Additional configuration options for a redis client
+/// Additional configuration options for a redis client.
 #[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
 pub struct RedisConfigOptions {
-    /// Maximum number of connections managed by the pool
+    /// Maximum number of connections managed by the pool.
     #[serde(default = "default_max_connections")]
     pub max_connections: u32,
-
-    /// If true, the health of a connection will be verified before it's checked out of the pool
-    #[serde(skip, default)]
-    pub test_on_check_out: bool,
 }
 
 impl Default for RedisConfigOptions {
     fn default() -> Self {
         Self {
             max_connections: default_max_connections(),
-            test_on_check_out: bool::default(),
         }
     }
 }
@@ -36,7 +31,7 @@ pub enum RedisConfig {
         /// This can also be a single node which is configured in cluster mode.
         cluster_nodes: Vec<String>,
 
-        /// Additional configuration options for the redis client and a connections pool
+        /// Additional configuration options for the redis client and a connections pool.
         #[serde(flatten)]
         options: RedisConfigOptions,
     },
@@ -46,15 +41,76 @@ pub enum RedisConfig {
     /// Contains the `redis://` url to the node.
     Single(String),
 
-    /// Connect to a single Redis instance
+    /// Connect to a single Redis instance.
     ///
-    /// Allows to provide more configuration options, e.g. `max_connections`
+    /// Allows to provide more configuration options, e.g. `max_connections`.
     SingleWithOpts {
-        /// Containes the `redis://` url to the node
+        /// Containes the `redis://` url to the node.
         server: String,
 
-        /// Additional configuration options for the redis client and a connections pool
+        /// Additional configuration options for the redis client and a connections pool.
         #[serde(flatten)]
         options: RedisConfigOptions,
     },
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_redis_single_opts() {
+        let yaml = r###"
+server: "redis://127.0.0.1:6379"
+max_connections: 42
+"###;
+
+        let config: RedisConfig = serde_yaml::from_str(yaml)
+            .expect("Parsed processing redis config: single with options");
+
+        match config {
+            RedisConfig::SingleWithOpts { server, options } => {
+                assert_eq!(options.max_connections, 42);
+                assert_eq!(server, "redis://127.0.0.1:6379");
+            }
+            e => panic!("Expected RedisConfig::SingleWithOpts but got {:?}", e),
+        }
+    }
+
+    #[test]
+    fn test_redis_single_opts_default() {
+        let yaml = r###"
+server: "redis://127.0.0.1:6379"
+"###;
+
+        let config: RedisConfig = serde_yaml::from_str(yaml)
+            .expect("Parsed processing redis config: single with options");
+
+        match config {
+            RedisConfig::SingleWithOpts { options, .. } => {
+                // check if all the defaults are correctly set
+                assert_eq!(options.max_connections, 24);
+            }
+            e => panic!("Expected RedisConfig::SingleWithOpts but got {:?}", e),
+        }
+    }
+
+    // To make sure that we have backwards compatibility and still support the redis configuration
+    // when the single `redis://...` address is provided.
+    #[test]
+    fn test_redis_single() {
+        let yaml = r###"
+"redis://127.0.0.1:6379"
+"###;
+
+        let config: RedisConfig = serde_yaml::from_str(yaml)
+            .expect("Parsed processing redis config: single with options");
+
+        match config {
+            RedisConfig::Single(server) => {
+                assert_eq!(server, "redis://127.0.0.1:6379");
+            }
+            e => panic!("Expected RedisConfig::Single but got {:?}", e),
+        }
+    }
 }
